fix(reader): extract PageDeepZoomButton — self-contained state + overlay
The inline approach (deepZoomOpen state on TranslationEditor + a root-level
conditional overlay) did not work in-browser: the button's onClick fired but
the overlay never mounted (verified via the live React tree — handler wired
correctly, page.deepzoom present, yet no overlay/canvas/tile requests; one
error-boundary 'Error' logged). A React-Compiler interaction with the
2400-line component's inline-IIFE layout + hoisted conditional.

Fix: move the open-state and the lazily-imported overlay into a small
self-contained PageDeepZoomButton — the same simple shape that works on the
artwork hero. Mounted once per reader layout; each instance is independent.
